chore(core): fix crate boundary leaks, visibility gaps, and duplicates (#1034)
